split experimental rotation out and prepare gargoyle for delay
