+++ sim/deathknight/dps/rotation_unholy.go
@@ -45,10 +45,7 @@ func (dk *DpsDeathknight) setupUnholySsOpener() {
 		NewAction(dk.RotationActionCallback_Garg).
 		NewAction(dk.RotationAction_CancelBT).
 		NewAction(dk.RotationActionCallback_ERW).
-		NewAction(dk.RotationActionCallback_BP).
-		NewAction(dk.RotationActionCallback_SS).
-		NewAction(dk.RotationActionCallback_SS).
-		NewAction(dk.getBloodRuneAction(false))
+		NewAction(dk.RotationActionCallback_BP)
 
 	dk.Main.NewAction(dk.RotationActionCallback_UnholySsRotation)
 }
@@ -65,8 +62,7 @@ func (dk *DpsDeathknight) setupUnholySsArmyOpener() {
 		NewAction(dk.RotationAction_CancelBT).
 		NewAction(dk.RotationActionCallback_ERW).
 		NewAction(dk.RotationActionCallback_AOTD).
-		NewAction(dk.RotationActionCallback_BP).
-		NewAction(dk.RotationActionCallback_SS)
+		NewAction(dk.RotationActionCallback_BP)
 
 	dk.Main.NewAction(dk.RotationActionCallback_UnholySsRotation)
 }
@@ -76,109 +72,35 @@ func (dk *DpsDeathknight) setupUnholyDndOpener() {
 		NewAction(dk.getFirstDiseaseAction()).
 		NewAction(dk.getSecondDiseaseAction()).
 		NewAction(dk.getBloodRuneAction(true)).
-		NewAction(dk.RotationActionCallback_DND).
-		NewAction(dk.RotationActionCallback_BT).
-		NewAction(dk.RotationActionCallback_UP).
-		NewAction(dk.RotationActionCallback_Garg).
-		NewAction(dk.RotationAction_CancelBT).
-		NewAction(dk.RotationActionCallback_ERW).
-		NewAction(dk.RotationActionCallback_BP).
-		NewAction(dk.RotationActionCallback_SS).
-		NewAction(dk.RotationActionCallback_SS).
-		NewAction(dk.getBloodRuneAction(false))
+		NewAction(dk.RotationActionCallback_DND)
 
 	if dk.Rotation.DeathAndDecayPrio == proto.Deathknight_Rotation_MaxRuneDowntime {
-		dk.Main.NewAction(dk.RotationActionCallback_UnholyDndRotation)
+		dk.Main.Clear().NewAction(dk.RotationActionCallback_UnholyDndRotation)
 	} else {
+		// Static opener with no Proc checks for gargoyle
+		dk.Opener.
+			NewAction(dk.RotationActionCallback_BT).
+			NewAction(dk.RotationActionCallback_UP).
+			NewAction(dk.RotationActionCallback_Garg).
+			NewAction(dk.RotationAction_CancelBT).
+			NewAction(dk.RotationActionCallback_ERW).
+			NewAction(dk.RotationActionCallback_BP)
+
+		// Experimental rotation with sequences
 		dk.dndStartSequence()
 	}
-}
-
-func (dk *DpsDeathknight) dndStartSequence() {
-	dk.Main.Clear().NewAction(dk.getFirstDiseaseAction()).
-		NewAction(dk.getSecondDiseaseAction()).
-		NewAction(dk.getBloodRuneAction(true)).
-		NewAction(dk.RotationAction_Dnd_Custom).
-		NewAction(dk.RotationAction_UnholyDndRotationGhoulFrenzyCheck)
-}
-
-// Custom Dnd callback with delay
-func (dk *DpsDeathknight) RotationAction_Dnd_Custom(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
-	casted := dk.CastDeathAndDecay(sim, target)
-	if !casted {
-		if !dk.DeathAndDecay.CD.IsReady(sim) {
-			dk.WaitUntil(sim, dk.DeathAndDecay.ReadyAt())
-			return true
-		}
-	} else {
-		s.Advance()
-	}
-	return casted
-}
-
-func (dk *DpsDeathknight) RotationAction_UnholyDndRotationGhoulFrenzyCheck(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
-	dk.Main.Clear()
-
-	if dk.Talents.GhoulFrenzy && (!dk.GhoulFrenzyAura.IsActive() || dk.GhoulFrenzyAura.RemainingDuration(sim) < time.Second*10) {
-		if dk.ur.ffFirst {
-			dk.Main.NewAction(dk.RotationActionCallback_IT).
-				NewAction(dk.RotationActionCallback_GF)
-		} else {
-			dk.Main.NewAction(dk.RotationActionCallback_GF).
-				NewAction(dk.RotationActionCallback_IT)
-		}
-	} else {
-		if dk.Talents.ScourgeStrike {
-			dk.Main.NewAction(dk.RotationActionCallback_SS)
-		} else {
-			dk.Main.NewAction(dk.getFirstDiseaseAction()).
-				NewAction(dk.getSecondDiseaseAction())
-		}
-	}
-
-	if dk.desolationAuraCheck(sim) {
-		dk.Main.NewAction(dk.RotationActionCallback_BS)
-	} else {
-		dk.Main.NewAction(dk.RotationActionCallback_BB)
-	}
-	dk.Main.NewAction(dk.RotationAction_UnholyDndRotationEnd)
-
-	dk.WaitUntil(sim, sim.CurrentTime)
-	return true
-}
-
-func (dk *DpsDeathknight) RotationAction_UnholyDndRotationEnd(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
-	dk.dndStartSequence()
-	dk.WaitUntil(sim, sim.CurrentTime)
-	return true
 }
 
 func (dk *DpsDeathknight) RotationActionCallback_UnholyDndRotation(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
 	casted := false
 
+	if dk.uhGargoyleCheck(sim, target) {
+		return true
+	}
+
 	if dk.Talents.GhoulFrenzy && !dk.uhShouldWaitForDnD(sim, false, true, true) {
-		// If no Ghoul Frenzy Aura or duration less then 10 seconds we try recasting
-		if !dk.GhoulFrenzyAura.IsActive() || dk.GhoulFrenzyAura.RemainingDuration(sim) < 10*time.Second {
-			if dk.CanBloodTap(sim) && dk.GhoulFrenzy.IsReady(sim) && dk.AllBloodRunesSpent() && dk.AllUnholySpent() && dk.SummonGargoyle.CD.TimeToReady(sim) > time.Second*60 {
-				// Use Ghoul Frenzy with a Blood Tap and Blood rune if all blood runes are on CD and Garg wont come off cd in less then a minute.
-				// The gargoyle check is there because you should BT -> UP -> Garg (Not in the sim yet)
-				if dk.uhDiseaseCheck(sim, target, dk.GhoulFrenzy, true, 1) {
-					dk.ghoulFrenzySequence(sim, true)
-					return true
-				} else {
-					dk.recastDiseasesSequence(sim)
-					return true
-				}
-			} else if !dk.Rotation.BtGhoulFrenzy && dk.CanGhoulFrenzy(sim) && dk.CanIcyTouch(sim) {
-				// Use Ghoul Frenzy with an Unholy Rune and sync the frost rune with Icy Touch
-				if dk.uhDiseaseCheck(sim, target, dk.GhoulFrenzy, true, 5) && dk.uhDiseaseCheck(sim, target, dk.IcyTouch, true, 5) {
-					dk.ghoulFrenzySequence(sim, false)
-					return true
-				} else {
-					dk.recastDiseasesSequence(sim)
-					return true
-				}
-			}
+		if dk.uhGhoulFrenzyCheck(sim, target) {
+			return true
 		}
 	}
 
@@ -217,7 +139,9 @@ func (dk *DpsDeathknight) RotationActionCallback_UnholyDndRotation(sim *core.Sim
 					}
 				}
 				if !casted {
-					casted = dk.CastDeathCoil(sim, target)
+					if !(dk.SummonGargoyle.IsReady(sim) || dk.SummonGargoyle.CD.TimeToReady(sim) < 5*time.Second) || dk.CurrentRunicPower() >= 100 {
+						casted = dk.CastDeathCoil(sim, target)
+					}
 					if !casted {
 						casted = dk.CastHornOfWinter(sim, target)
 					}
@@ -226,46 +150,48 @@ func (dk *DpsDeathknight) RotationActionCallback_UnholyDndRotation(sim *core.Sim
 		}
 	}
 
-	return casted
-}
+	// Gargoyle cast needs to be checked more often then default rotation on gcd/resource gain checks
+	if dk.SummonGargoyle.IsReady(sim) {
+		dk.WaitUntil(sim, sim.CurrentTime+100*time.Millisecond)
+		return true
+	}
 
-func (dk *DpsDeathknight) desolationAuraCheck(sim *core.Simulation) bool {
-	return !dk.DesolationAura.IsActive() || dk.DesolationAura.RemainingDuration(sim) < 10*time.Second || dk.Env.GetNumTargets() == 1
+	return casted
 }
 
 func (dk *DpsDeathknight) RotationActionCallback_UnholySsRotation(sim *core.Simulation, target *core.Unit, s *deathknight.Sequence) bool {
 	casted := false
 
+	if dk.uhGargoyleCanCast(sim) {
+		dk.CastBloodTap(sim, dk.CurrentTarget)
+		dk.CastUnholyPresence(sim, dk.CurrentTarget)
+		casted = dk.CastSummonGargoyle(sim, target)
+		if casted {
+			return true
+		}
+	}
+
+	// Go back to Blood Presence after gargoyle cast
+	if dk.PresenceMatches(deathknight.UnholyPresence) && !dk.CanSummonGargoyle(sim) {
+		if dk.BloodTapAura.IsActive() {
+			dk.BloodTapAura.Deactivate(sim)
+		}
+		if dk.CastBloodPresence(sim, target) {
+			dk.WaitUntil(sim, sim.CurrentTime)
+			return true
+		}
+	}
+
 	if dk.Talents.GhoulFrenzy {
-		// If no Ghoul Frenzy Aura or duration less then 10 seconds we try recasting
-		if !dk.GhoulFrenzyAura.IsActive() || dk.GhoulFrenzyAura.RemainingDuration(sim) < 10*time.Second {
-			if dk.CanBloodTap(sim) && dk.GhoulFrenzy.IsReady(sim) && dk.AllBloodRunesSpent() && dk.AllUnholySpent() && dk.SummonGargoyle.CD.TimeToReady(sim) > time.Second*60 {
-				// Use Ghoul Frenzy with a Blood Tap and Blood rune if all blood runes are on CD and Garg wont come off cd in less then a minute.
-				// The gargoyle check is there because you should BT -> UP -> Garg (Not in the sim yet)
-				if dk.uhDiseaseCheck(sim, target, dk.GhoulFrenzy, true, 1) {
-					dk.ghoulFrenzySequence(sim, true)
-					return true
-				} else {
-					dk.recastDiseasesSequence(sim)
-					return true
-				}
-			} else if !dk.Rotation.BtGhoulFrenzy && dk.CanGhoulFrenzy(sim) && dk.CanIcyTouch(sim) {
-				// Use Ghoul Frenzy with an Unholy Rune and sync the frost rune with Icy Touch
-				if dk.uhDiseaseCheck(sim, target, dk.GhoulFrenzy, true, 5) && dk.uhDiseaseCheck(sim, target, dk.IcyTouch, true, 5) {
-					dk.ghoulFrenzySequence(sim, false)
-					return true
-				} else {
-					dk.recastDiseasesSequence(sim)
-					return true
-				}
-			}
+		if dk.uhGhoulFrenzyCheck(sim, target) {
+			return true
 		}
 	}
 
 	// What follows is a simple APL where every cast is checked against current diseses
 	// And if the cast would leave the DK with not enough runes to cast disease before falloff
 	// the cast is canceled and a disease recast is queued. Priority is as follows:
-	// Scourge Strike -> Blood Strike (or Pesti on Aoe) -> Death Coil -> Horn of Winter
+	// Scourge Strike -> Blood Strike (or Pesti/BB on Aoe) -> Death Coil -> Horn of Winter
 	if !casted {
 		if dk.uhDiseaseCheck(sim, target, dk.ScourgeStrike, true, 1) {
 			casted = dk.CastScourgeStrike(sim, target)
@@ -278,10 +204,10 @@ func (dk *DpsDeathknight) RotationActionCallback_UnholySsRotation(sim *core.Simu
 				casted = dk.uhSpreadDiseases(sim, target, s)
 			} else {
 				if dk.uhDiseaseCheck(sim, target, dk.BloodStrike, true, 1) {
-					if dk.Env.GetNumTargets() > 1 && dk.DesolationAura.IsActive() && dk.DesolationAura.RemainingDuration(sim) > time.Second*10 {
-						casted = dk.CastBloodBoil(sim, target)
-					} else {
+					if dk.desolationAuraCheck(sim) {
 						casted = dk.CastBloodStrike(sim, target)
+					} else {
+						casted = dk.CastBloodBoil(sim, target)
 					}
 				} else {
 					dk.recastDiseasesSequence(sim)
@@ -295,6 +221,12 @@ func (dk *DpsDeathknight) RotationActionCallback_UnholySsRotation(sim *core.Simu
 				}
 			}
 		}
+	}
+
+	// Gargoyle cast needs to be checked more often then default rotation on gcd/resource gain checks
+	if dk.SummonGargoyle.IsReady(sim) {
+		dk.WaitUntil(sim, sim.CurrentTime+100*time.Millisecond)
+		return true
 	}
 
 	return casted
